ai_training/requirements.txt
openai>=1.0.0
python-dotenv>=1.0.0
sounddevice>=0.4.6
soundfile>=0.12.1
numpy>=1.24.0
flask>=3.0.0
flask-cors>=4.0.0
pydub>=0.25.1
